Make user_id field nullable in payment entity, return a double number as a current rate

+++ backend/src/main/java/io/github/luminaire1337/propertyvista/backend/dto/response/PaymentRateResponse.java
@@ -1,6 +1,6 @@
 package io.github.luminaire1337.propertyvista.backend.dto.response;
 
 public record PaymentRateResponse(
-        Integer currentRate
+        Double currentRate
 ) {
 }
